services/assistant.py
import sys
import os
from llama_cpp import Llama
import pandas as pd
from database.analyzer import DatabaseAnalyzer
from database.validator import QueryValidator

class LlamaSQLAssistant:
    def __init__(self, model_path: str, db_analyzer: DatabaseAnalyzer):
        self.model_path = model_path
        self.db_analyzer = db_analyzer
        self.validator = QueryValidator()
        # Initialize Llama model here
        self.llama_model = Llama(model_path=self.model_path)
        
    def generate_sql(self, user_query: str) -> str:
        # Implement SQL generation logic
        return f"SELECT * FROM table WHERE condition = 'value' LIMIT 10;"
    
    def execute_query(self, sql: str) -> pd.DataFrame:
        return pd.read_sql_query(sql, self.db_analyzer.conn)