requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
pydantic==1.8.2
sqlalchemy==1.4.23
pandas==1.3.3
redis==4.0.2
llama-cpp-python==0.1.0
sqlparse==0.4.2
python-jose==3.3.0